+++ db/error.go
@@ -3,7 +3,7 @@ package db
 import (
 	"errors"
 
-	"github.com/jackc/pgx/v5/pgconn"
+	"github.com/jackc/pgconn"
 	"github.com/lib/pq"
 )